// eshop-libs/projects/sag-article-detail/src/enums/availability-info.enum.ts
export enum AVAILABILITY_INFO {
    PICKUP = 'PICKUP',
    TOUR = 'TOUR',
    UNIT_LOAD = 'UNIT_LOAD',
    EXPRESS = 'EXPRESS',
    POST_NORMAL = 'POST_NORMAL',
    SERVICE_TECHNICAN = 'SERVICE_TECHNICAN',
    NES_NIGHT = 'NES_NIGHT',
    NIGHTLY_DELIVERY = 'NIGHTLY_DELIVERY',
    NES_DAY = 'NES_DAY',
    WAREHOUSE_TIRES_DELIVERY = 'WAREHOUSE_TIRES_DELIVERY',
    DIRECT_TIRES_DELIVERY = 'DIRECT_TIRES_DELIVERY',
    NOT_AVAILABILITY = 'TO BE CLARIFY',
    AVAIL_STATE_INORDER_24_HOURS = 10,
    AVAIL_STATE_NO_STOCK = 11,
    IS_SOFORT_STATE = 1,
    IS_TOUR_STATE = 2,
    IS_ARTICLE_24H_STATE = 3
}

export enum COLLECTIVE_DELIVERY {
    COLLECTIVE_DELIVERY1 = 'COLLECTIVE_DELIVERY1'
}

export enum CZ_AVAIL_STATE {
    IN_STOCK_BRANCH = 2,
    PARTLY_AVAILABLE = 3,
    UNAVAILABLE = 4,
    IN_STOCK_CENTRAL_WAREHOUSE = 5,
    NON_ORDERALE = 9
}

export enum SB_LOCATION_STATE {
    GREEN = 'GREEN',
    YELLOW = 'YELLOW',
    RED = 'RED',
    BLUE = 'BLUE'
}

export enum SB_LOCATION_NAME {
    MULTI_SHOP = 'MULTI_SHOP',
    FULLY_EXTERNAL = 'FULLY_EXTERNAL',
    PARTLY_EXTERNAL = 'PARTLY_EXTERNAL',
    PARTLY_LOCAL = 'PARTLY_LOCAL',
    NONE = 'NONE',
    NOT_AVAILABLE = 'NOT_AVAILABLE'
}

export enum SB_LOCATION_TYPES {
    NON_PRIORITY = 'NON_PRIORITY',
    PRIMARY = 'PRIMARY',
    SECONDARY = 'SECONDARY',
    THIRD = 'THIRD'
}